pr reviews: renamed files and directories, added namespaces, added projects to solution
