+++ Fhir.Liquid.Converter.sln
@@ -22,6 +22,14 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Health.Fhir.TemplateManagement.UnitTests", "src\Microsoft.Health.Fhir.TemplateManagement.UnitTests\Microsoft.Health.Fhir.TemplateManagement.UnitTests.csproj", "{C4049E7B-7977-4691-BE86-4BD7E3E8C401}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{56C74465-CD00-48C1-9A89-189FAFD44308}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dibbs.FhirConverterApi", "src\Dibbs.FhirConverterApi\Dibbs.FhirConverterApi.csproj", "{4CE994FB-1614-40F7-9FED-0314F8B02C04}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dibbs.FhirConverterApi.UnitTests", "src\Dibbs.FhirConverterApi.UnitTests\Dibbs.FhirConverterApi.UnitTests.csproj", "{74C7362F-23E3-4C7A-A95D-5C0DBEEDCBD6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Dibbs.FhirConverterApi.FunctionalTests", "src\Dibbs.FhirConverterApi.FunctionalTests\Dibbs.FhirConverterApi.FunctionalTests.csproj", "{53583FEF-D7B1-4781-AF7A-98F59D3E477F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -56,11 +64,28 @@ Global
 		{C4049E7B-7977-4691-BE86-4BD7E3E8C401}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C4049E7B-7977-4691-BE86-4BD7E3E8C401}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C4049E7B-7977-4691-BE86-4BD7E3E8C401}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4CE994FB-1614-40F7-9FED-0314F8B02C04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CE994FB-1614-40F7-9FED-0314F8B02C04}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CE994FB-1614-40F7-9FED-0314F8B02C04}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CE994FB-1614-40F7-9FED-0314F8B02C04}.Release|Any CPU.Build.0 = Release|Any CPU
+		{74C7362F-23E3-4C7A-A95D-5C0DBEEDCBD6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{74C7362F-23E3-4C7A-A95D-5C0DBEEDCBD6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{74C7362F-23E3-4C7A-A95D-5C0DBEEDCBD6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{74C7362F-23E3-4C7A-A95D-5C0DBEEDCBD6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{53583FEF-D7B1-4781-AF7A-98F59D3E477F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{53583FEF-D7B1-4781-AF7A-98F59D3E477F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{53583FEF-D7B1-4781-AF7A-98F59D3E477F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{53583FEF-D7B1-4781-AF7A-98F59D3E477F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DF59726B-5E9A-43A5-B927-F2ECF15508AC}
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{4CE994FB-1614-40F7-9FED-0314F8B02C04} = {56C74465-CD00-48C1-9A89-189FAFD44308}
+		{74C7362F-23E3-4C7A-A95D-5C0DBEEDCBD6} = {56C74465-CD00-48C1-9A89-189FAFD44308}
+		{53583FEF-D7B1-4781-AF7A-98F59D3E477F} = {56C74465-CD00-48C1-9A89-189FAFD44308}
 	EndGlobalSection
 EndGlobal
