+++ src/Dibbs.FhirConverterApi.UnitTests/Dibbs.FhirConverterApi.UnitTests.csproj
@@ -25,8 +25,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Microsoft.Health.Fhir.Liquid.Converter.FunctionalTests\Microsoft.Health.Fhir.Liquid.Converter.FunctionalTests.csproj" />
-    <ProjectReference Include="..\FHIRConverterAPI\FHIRConverterAPI.csproj" />
+    <ProjectReference Include="..\Dibbs.FhirConverterApi\Dibbs.FhirConverterApi.csproj" />
   </ItemGroup>
 
 </Project>
