+++ src/Dibbs.FhirConverterApi.UnitTests/TestData/FileSingleMessageInvalidSegments.hl7
