+++ src/Dibbs.FhirConverterApi.UnitTests/unit/ConverterLogicHandlerTest.cs
@@ -1,3 +1,5 @@
+using Dibbs.FhirConverterApi;
+
 public class ConverterLogicHandlerTest
 {
     [Fact]
