+++ src/Dibbs.FhirConverterApi/Models/UserFacingException.cs
@@ -1,0 +1,23 @@
+using System.Net;
+
+namespace Dibbs.FhirConverterApi.Models;
+
+/// <summary>
+/// Allows customization of exception message, specifying status code, and optionally includes the original exception.
+/// </summary>
+public class UserFacingException : Exception
+{
+    public UserFacingException(string message, HttpStatusCode statusCode)
+    : base(message)
+    {
+        StatusCode = statusCode;
+    }
+
+    public UserFacingException(string message, HttpStatusCode statusCode, Exception? innerException = null)
+    : base(message, innerException)
+    {
+        StatusCode = statusCode;
+    }
+
+    public HttpStatusCode StatusCode { get; set; }
+}
